demos/rag_min/requirements.txt
numpy>=1.24
rich>=13.7
sentence-transformers>=2.6
faiss-cpu>=1.7.4